requirements.txt
# Core dependencies
streamlit>=1.31.0
anthropic>=0.40.0
chromadb>=0.4.22
python-dotenv>=1.0.0

# For processing and NLP
numpy>=1.24.0

# PDF processing
pdfplumber>=0.10.0
